+++ tests/CMakeLists.txt
@@ -62,12 +62,7 @@ endif()
 
 if(LIBQUIC_BUILD_SPEEDTEST)
     set(LIBQUIC_SPEEDTEST_PREFIX "" CACHE STRING "Binary prefix for speedtest binaries")
-    set(speedtests speedtest-client speedtest-server dgram-speed-client dgram-speed-server ping-server)
-    if(NOT WIN32)
-        # Currently doesn't build on windows because it relies on sigset_t to handle stopping it.
-        set(speedtests ${speedtests} ping-client)
-    endif()
-    foreach(x ${speedtests})
+    foreach(x speedtest-client speedtest-server)
         add_executable(${x} ${x}.cpp)
         target_link_libraries(${x} PRIVATE tests_common)
         set_target_properties(${x} PROPERTIES OUTPUT_NAME ${LIBQUIC_SPEEDTEST_PREFIX}${x})
@@ -76,7 +71,7 @@ if(LIBQUIC_BUILD_SPEEDTEST)
 
     if(LIBQUIC_INSTALL)
         install(
-            TARGETS ${speedtests}
+            TARGETS speedtest-client speedtest-server
             DESTINATION ${CMAKE_INSTALL_BINDIR}
         )
     endif()
