+++ tests/utils.cpp
@@ -229,7 +229,7 @@ namespace oxen::quic
     void add_log_opts(CLI::App& cli, std::string& file, std::string& level)
     {
         file = "stderr";
-        level = "info";
+        level = "warning";
 
         cli.add_option("-l,--log-file", file, "Log output filename, or one of stdout/-/stderr/syslog.")
                 ->type_name("FILE")
